test: テストヘルパー集約とバリデーションテスト追加
